Support setting up Openvswitch (Ovs) Bridge network
Resolves: https://issues.redhat.com/browse/RHEL-33465

This patch supports setting up an Ovs bridge in kdump initrd. An Ovs
bridge is similar to a classic Linux bridge but we use ovs-vsctl to find
out the Ethernet device (having the MAC address as the bridge) added to
an Ovs bridge.

In theory, we can also try to bring up the original physical network
interface before ovs-configuration.service [1][2]. But this approach is
cumbersome because it breaks our assumption that we should bring up the
same work in kdump intrd as in 1st kernel.

[1] https://github.com/openshift/machine-config-operator/blob/master/templates/common/_base/units/ovs-configuration.service.yaml
[2] https://github.com/openshift/machine-config-operator/blob/master/templates/common/_base/files/configure-ovs-network.yaml

Signed-off-by: Coiby Xu <coxu@redhat.com>