kdumpctl: Rewrite load_kdump
This rewrite removes an unneeded global variable and the need to
manually keep the command and debug print in sync. In addition it allows
to easily check for unsupported options and print an error message. This
becomes important as the kexec_load system call (option -c) will be
disabled in the RHEL10 but not the Fedora kernel. So having this patch
allows to have a simple RHEL-only patch without deviating from upstream
too much.

Signed-off-by: Philipp Rudo <prudo@redhat.com>