tests: add Test Management Tool kdump NFS tests
Currently, all integration tests fail F40 because of failing to install
package and enable services on the cloud image properly. TMT (Test Management
Tool) supports running tests on virtual machines and provides an
user-friendly interface (YAML) to define tasks like installing packages,
running commands on the virtual machines. By using TMT, we no longer
need to spend time and efforts on maintaining the current test framework
and can also enjoy rich handy features like managing tests provided by TMT.

There are six test steps in TMT. Quoting [1], here is an overview,
 - discover
   Gather information about test cases to be executed.

 - provision
   Provision an environment for testing or use localhost.

 - prepare
   Prepare the environment for testing.

 - execute
   Run tests using the specified executor.

 - report
   Provide test results overview and send reports.

 - finish
   Perform the finishing tasks and clean up provisioned guests.

And each step step is configurable via the so-called plans [2]. TMT also
uses plans to group relevant tests and select needed tests.

This commit adds the test of sending crash dump to a NFS target by
setting two VMs using TMT's MultiHost Testing feature [3],
 - client
   Setup up NFS dump target and trigger crash
 - server
   acts as an NFS server and check sent vmcore

The scaffold of TMT test project can be created as follows,

    mkdir tests && cd tests
    tmt init --template mini
    mv plans/example/{example,kdump}.fmf

    tmt tests create /setup/default_crashkernel -t beakerlib
    tmt tests create /setup/nfs_server -t beakerlib
    tmt tests create /test/trigger_crash -t beakerlib
    tmt tests create /test/check_vmcore -t beakerlib

Then we fill in the content for created plans and tests,

    # tmt run -a
    /var/tmp/tmt/run-017

    /plans/kdump
        discover
            how: fmf
            name: client-setup
            directory: /root/kdump-tests
            tests: /setup/kdump
            how: fmf
            name: server-setup
    ...
            execute task #4: server-test on server
            how: tmt

            summary: 4 tests executed
        report
            how: display
            summary: 4 tests passed
        finish

A main.fmf is put under the root directory of tests and setup for other
tests to inherit [4].

When there is a kernel panic, ssh access to the guest will exit with
with 255 and TMT will retreat it as an error. We can use
restart-on-exit-code=255 [5] to let TMT this is expected.

[1] https://tmt.readthedocs.io/en/stable/overview.html#run
[2] https://tmt.readthedocs.io/en/stable/spec/plans.html
[3] https://tmt.readthedocs.io/en/stable/guide.html#multihost-testing
[4] https://tmt.readthedocs.io/en/stable/guide.html#inheritance
[5] https://tmt.readthedocs.io/en/latest/spec/tests.html#restart

Signed-off-by: Coiby Xu <coxu@redhat.com>