lib: Ensure we don't find bind mounts for device target
There's comment here that `--source` somehow avoids bind
mounts, but that appears not to be the case in my
testing. I think we just happened to be lucky before
now with the `--first` picking the value we wanted.

Instead of using `--first` and hoping for the best,
parse the mounts and skip ones which are bind mounts
explicitly.

Signed-off-by: Colin Walters <walters@verbum.org>
Reviewed-by: Lichen Liu <lichliu@redhat.com>