dracut: Disable ostree-prepare-root
In some images such as the recent fedora/rhel bootc base image,
the ostree dracut module is statically enabled:
https://gitlab.com/fedora/bootc/base-images/-/blob/40df0eb3827763039e4d17264968e37f850b70e5/tier-0/initramfs.yaml#L9

And also recently, we changed the ostree systemd unit
to enter emergency.target if it fails in:
https://github.com/ostreedev/ostree/pull/3220/commits/05b3b662756def30026fc722d0832a24a33d9803

These two things combine mean we'll fail before kdump gets
a chance to run.

For our use case we don't need ostree in the initrd.

I tried to override this with `--omit=ostree` in our dracut
invocation, but that causes an error (dracut doesn't let the
cmdline override static config).

For now, let's just mask the service in our initrd.

Signed-off-by: Colin Walters <walters@verbum.org>
Reviewed-by: Lichen Liu <lichliu@redhat.com>