editorconfig: set indent_size=8
The indent_size gives the number of spaces an indentation should have.
When combined with indent_style=tab and tab_width=X the leading X spaces
are replaced by a tab. If tab_width is omitted it is set to the value of
indent_size. For the current settings this means that every level of
indentation is indented by a single space which gets replaced by a tab.
The tab again is displayed by a single space, if you set your editor to
follow the editconfig. This is not only barely readable but also impacts
line breaks for over long lines. Thus set indent_size=8 to better match
the way over long lines are broken at the moment.

Signed-off-by: Philipp Rudo <prudo@redhat.com>